Passed Wave 6 (some refactoring)
