+++ app/__init__.py
@@ -29,7 +29,8 @@ def create_app(test_config=None):
     db.init_app(app)
     migrate.init_app(app, db)
 
-    from .routes import tasks_bp, goals_bp
+    from .task_routes import tasks_bp
+    from .goal_routes import goals_bp
     app.register_blueprint(tasks_bp)
     app.register_blueprint(goals_bp)
